Collection renamed to SimpleCollection, Added exclusions to the collection relation so the collection works with groups. Added relations parameter to the constructor to provide vustome relations
